Fix for issue #109 Support for Asus U36JC

+++ install.sh
@@ -37,7 +37,7 @@
 #    You should have received a copy of the GNU General Public License
 #    along with bumblebee.  If not, see <http://www.gnu.org/licenses/>.
 #
-BUMBLEBEEVERSION=1.4.27
+BUMBLEBEEVERSION=1.4.28
 
 
 ROOT_UID=0
@@ -257,6 +257,7 @@ case "$DISTRO" in
    ;;
    4)
     echo "Skip drivers installation. Please remember that NVidia drivers *HAVE TO BE INSTALLED*"
+
    ;;
    *)
     echo
@@ -299,35 +300,25 @@ echo
 cp install-files/xorg.conf.intel /etc/X11/xorg.conf
 cp install-files/xorg.conf.nvidia /etc/X11/
 
-case `diff install-files/bumblebee-enablecard /usr/local/bin/bumblebee-enablecard` in 
+if [ ! -f install-files/bumblebee-enablecard ]; then
+ # Not installed
+ cp install-files/bumblebee-enablecard /usr/local/bin/
+else
+ # Already Exists
+ echo
+ echo "nVidia card enable-script: /usr/local/bin/bumblebee-enablecard, already exists not overwriting"
+ echo
+fi
 
-0) # No differens
-;;
-
-1) # Differs
-echo "nVidia card enable-script: /usr/local/bin/bumblebee-enablecard, has been modified, not overwriting"
-;;
-
-2) # Not installed 
- cp install-files/bumblebee-enablecard /usr/local/bin/ 
-;;
-
-esac
-
-case `diff install-files/bumblebee-disablecard /usr/local/bin/bumblebee-disablecard` in          
-
-0) # No differens
-;;
-
-1) # Differs
-echo "nVidia card disable-script: /usr/local/bin/bumblebee-disablecard, has been modified, not overwriting"
-;;
-  
-2) # Not installed 
+if [ ! -f install-files/bumblebee-disablecard ]; then
+ # Not installed
  cp install-files/bumblebee-disablecard /usr/local/bin/
-;;
-     
-esac
+else
+ # Already Exists
+ echo
+ echo "nVidia card disable-script: /usr/local/bin/bumblebee-disablecard, already exists not overwriting"    
+ echo
+fi
 
 cp -n $BASHRC $BASHRC.optiorig
 
@@ -560,7 +551,7 @@ echo "6) Dell Vostro 3300"
 echo "7) Dell Vostro 3400/3500"
 echo "8) Samsung RF511/RF711/QX410-J01"
 echo "9) Toshiba Satellite M645-SP4132L"
-echo "10) Asus U30J/U35J/U43JC/U35JC/U43JC/U53JC/P52JC/K52JC/X52JC/N53SV/N61JV/X64JV"
+echo "10) Asus U30J/U35J/U36JC/U43JC/U35JC/U43JC/U53JC/P52JC/K52JC/X52JC/N53SV/N61JV/X64JV"
 echo
 echo "97) Manually Set Output to CRT-0"
 echo "98) Manually Set Output to DFP-0"
